+++ frontend/src/components/member/ChattingMember.vue
@@ -2,15 +2,20 @@
   <div>
     <!-- 서버로부터 받은 메시지를 표시 -->
     <div v-for="(message, index) in messages" :key="index">
-      <strong>({{ message.userId }}):</strong> {{ message.text }}
+      <strong>{{ message.from }}:</strong> {{ message.text }}
     </div>
 
-    <!-- 사용자 입력을 받는 부분 -->
-    <input v-model="messageText" @keyup.enter="sendMessage" placeholder="Enter message" />
+    <!-- 사용자 ID와 입력 필드를 함께 표시 -->
+    <div class="input-container">
+      <span class="user-id">{{ userId }}</span>  <!-- 사용자 ID 표시 -->
+      <input v-model="messageText" @keyup.enter="sendMessage" placeholder="Enter message" />
+    </div>
   </div>
 </template>
 
 <script>
+import { getUserIdFromToken } from  "../../apis/axios";  // api.js 파일 경로에 맞게 수정
+
 export default {
   data() {
     return {
@@ -18,10 +23,13 @@ export default {
       messageText: '',  // 사용자가 입력한 메시지
       messages: [],  // 서버에서 받은 메시지들을 저장
       username: 'Admin', // 사용자 이름
+      userId: null,  // 사용자 ID
     };
   },
   created() {
     this.connect();  // 컴포넌트 생성 시 WebSocket 연결
+    this.userId = getUserIdFromToken();  // 토큰에서 사용자 ID 추출
+    console.log("사용자 ID:", this.userId);  // 디코딩 결과 출력
   },
   methods: {
     // WebSocket 연결 설정
@@ -33,20 +41,21 @@ export default {
       };
       // 서버로부터 메시지를 수신했을 때 처리
       this.socket.onmessage = (event) => {
-  const message = JSON.parse(event.data);
-  
-  // message.text가 또 다른 JSON 문자열이라면 추가 파싱
-  let parsedText;
-  try {
-    parsedText = JSON.parse(message.text);
-  } catch (e) {
-    parsedText = message.text; // JSON 파싱이 실패하면 그냥 문자열 유지
-  }
-  console.log("여기왔음");
-  console.log("보내준아이디"+message.userId);
-  this.messages.push({ from: message.from, userId: message.userId, text: parsedText.text || parsedText });
-  console.log("보내준아이디"+message.userId);
-};
+        const message = JSON.parse(event.data);
+
+        // message.text가 JSON 형식인지 확인하고 추가 파싱
+        let parsedText = message.text;
+        if (typeof parsedText === "string") {
+          try {
+            const innerMessage = JSON.parse(parsedText);
+            parsedText = innerMessage.text || innerMessage; // 내부 메시지의 text 필드 사용
+          } catch (e) {
+            // JSON 파싱 실패 시 원본 유지
+          }
+        }
+
+        this.messages.push({ from: message.from, text: parsedText });
+      };
 
       this.socket.onerror = (error) => {
         console.error('WebSocket error: ', error);
@@ -62,6 +71,7 @@ export default {
         const message = {
           from: this.username,
           text: this.messageText,
+          userId: this.userId,
         };
         this.socket.send(JSON.stringify(message));  // JSON 형식으로 메시지 전송
         console.log('Message sent:', message);
@@ -84,6 +94,18 @@ export default {
 div {
   margin-bottom: 15px;
 }
+
+.input-container {
+  display: flex;
+  align-items: center;
+}
+
+.user-id {
+  margin-right: 10px;
+  font-weight: bold;
+  color: #333;
+}
+
 input {
   width: 100%;
   padding: 8px;
